pyproject.toml
[project]
name = "mcp-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain-anthropic>=0.3.20",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
]
